files_clustering/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
scikit-learn
numpy
pandas
fastapi
uvicorn
rapidfuzz
